fix: address pre-merge and post-merge review feedback
- fix(retryableLazy): change resetLabel from 'Reload Page' to 'Retry Load'
  to avoid duplicate 'Reload Page' buttons in ErrorBoundary fallback UI
- fix(Questionnaire): add onReset={resetSurvey} to ErrorBoundary wrapping
  RecommendationResults to prevent crash loop on scoring errors
- refactor(RecommendationResults): extract normalizeUserAnswers,
  calculateDeviceScores, and devices to src/utils/deviceScoring.tsx
  to resolve React Fast Refresh lint warnings from non-component exports
- refactor(QuestionnaireContext): replace hardcoded totalSteps=10 with
  dynamic setTotalSteps driven by questions array length, eliminating
  manual cross-module synchronization

All 72 tests pass. Lint clean.
