+++ app/.antigravitycli/d2c01658-51f9-4cf0-af7a-d270f373e3ae.json
@@ -1,0 +1,1 @@
+/home/luandro/.gemini/config/projects/d2c01658-51f9-4cf0-af7a-d270f373e3ae.json
